+++ src/ChessDotCpp/search.cpp
@@ -6,38 +6,114 @@
 #include "evaluation.h"
 #include "stopper.h"
 
-static void StoreTranspositionTable(SearchData& data, const ZobristKey key, const Move move, const Ply depth, const Score score, const TtFlag flag)
+void Search::StoreTranspositionTable(const ZobristKey key, const Move move, const Ply depth, const Score score, const TtFlag flag)
 {
-    if (data.Stopper.Stopped)
+    if (State.Stopper.Stopped)
     {
         return;
     }
     
-    data.Global.Table.Store(key, move, depth, score, flag);
+    State.Global.Table.Store(key, move, depth, score, flag);
 }
 
-static Score Contempt(const Board& board)
+Score Search::Contempt(const Board& board)
 {
     return 0;
 }
 
-static Score AlphaBeta(Board& board, SearchData& data, Ply depth, const Ply ply, Score alpha, Score beta)
+Score Search::Quiescence(Board& board, Ply depth, Ply ply, Score alpha, Score beta)
 {
-    //if (depth > 2 && data.Stopper.ShouldStop())
-    //{
-    //    const Score score = Contempt(board);
-    //    return score;
-    //}
+    ++State.Stats.Nodes;
+    
+    EachColor<Bitboard> pins;
+    PinDetector::GetPinnedToKings(board, pins);
+    const Score standPat = Evaluation::Evaluate(board, pins);
+
+    if (standPat >= beta)
+    {
+        return beta;
+    }
+
+    if (alpha < standPat)
+    {
+        alpha = standPat;
+    }
+
+    const Bitboard checkers = AttacksGenerator::GetCheckers(board);
+    const bool inCheck = checkers != BitboardConstants::Empty;
+    const Score currentMateScore = Constants::Mate - ply;
+
+    const Bitboard pinned = pins[board.ColorToMove];
+
+    MoveArray moves;
+    MoveCount moveCount = 0;
+    MoveGenerator::GetAllPotentialCaptures(board, checkers, pinned, moves, moveCount);
+
+    Score bestScore = -Constants::Inf;
+    Move bestMove;
+    bool raisedAlpha = false;
+    bool betaCutoff = false;
+    uint8_t movesEvaluated = 0;
+    for (MoveCount i = 0; i < moveCount; i++)
+    {
+        const Move move = moves[i];
+        const bool valid = MoveValidator::IsKingSafeAfterMove2(board, move, checkers, pinned);
+        if (!valid)
+        {
+            continue;
+        }
+
+        board.DoMove(move);
+        const Score childScore = -Quiescence(board, depth - 1, ply + 1, -beta, -alpha);
+        board.UndoMove();
+        movesEvaluated++;
+
+        if (childScore > bestScore)
+        {
+            bestScore = childScore;
+            bestMove = move;
+
+            if (childScore > alpha)
+            {
+                alpha = childScore;
+                raisedAlpha = true;
+
+                if (childScore >= beta)
+                {
+                    betaCutoff = true;
+                    break;
+                }
+            }
+        }
+    }
+
+    if (betaCutoff)
+    {
+        return beta;
+    }
+
+    return alpha;
+}
+
+Score Search::AlphaBeta(Board& board, Ply depth, const Ply ply, Score alpha, Score beta)
+{
+    if (depth > 2 && State.Stopper.ShouldStop())
+    {
+        const Score score = Contempt(board);
+        return score;
+    }
     
     if(depth <= 0)
     {
-        ++data.Stats.Nodes;
         
         EachColor<Bitboard> pins;
         PinDetector::GetPinnedToKings(board, pins);
-        const Score eval = Evaluation::Evaluate(board, pins);
+        //const Score eval = Evaluation::Evaluate(board, pins);
+        const Score eval = Quiescence(board, depth, ply, alpha, beta);
         return eval;
     }
+
+    ++State.Stats.Nodes;
 
     const Bitboard checkers = AttacksGenerator::GetCheckers(board);
     const bool inCheck = checkers != BitboardConstants::Empty;
@@ -67,7 +143,7 @@ static Score AlphaBeta(Board& board, SearchData& data, Ply depth, const Ply ply,
         }
 
         board.DoMove(move);
-        const Score childScore = -AlphaBeta(board, data, depth - 1, ply + 1, -beta, -alpha);
+        const Score childScore = -AlphaBeta(board, depth - 1, ply + 1, -beta, -alpha);
         board.UndoMove();
         movesEvaluated++;
         
@@ -93,7 +169,7 @@ static Score AlphaBeta(Board& board, SearchData& data, Ply depth, const Ply ply,
     
     if (betaCutoff)
     {
-        StoreTranspositionTable(data, board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Beta);
+        StoreTranspositionTable(board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Beta);
         return beta;
     }
 
@@ -109,58 +185,56 @@ static Score AlphaBeta(Board& board, SearchData& data, Ply depth, const Ply ply,
 
     if(raisedAlpha)
     {
-        StoreTranspositionTable(data, board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Exact);
+        StoreTranspositionTable(board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Exact);
     }
     else
     {
-        StoreTranspositionTable(data, board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Alpha);
+        StoreTranspositionTable(board.Key, bestMove, depth, bestScore, TranspositionTableFlags::Alpha);
     }
     
     return alpha;
 }
 
-static Score Aspiration(Board& board, SearchData& data, const Ply depth, const Score previous)
+Score Search::Aspiration(Board& board, const Ply depth, const Score previous)
 {
-    Score score = AlphaBeta(board, data, depth, 0, -Constants::Inf, Constants::Inf);
+    Score score = AlphaBeta(board, depth, 0, -Constants::Inf, Constants::Inf);
     return score;
 }
 
-static Move IterativeDeepen(Board& board, SearchData& data, const SearchCallback& callback)
+Move Search::IterativeDeepen(Board& board)
 {
-    Score score = AlphaBeta(board, data, 1, 0, -Constants::Inf, Constants::Inf);
-    data.Global.Table.SavePrincipalVariation(board);
-    SearchCallbackData callbackData(board, data, 1, score);
+    Score score = AlphaBeta(board, 1, 0, -Constants::Inf, Constants::Inf);
+    State.Global.Table.SavePrincipalVariation(board);
+    SearchCallbackData callbackData(board, State, 1, score);
     
-    callback(callbackData);
+    Callback(callbackData);
 
-    if (data.Stopper.ShouldStopDepthIncrease())
+    if (State.Stopper.ShouldStopDepthIncrease())
     {
-        return data.Global.Table.SavedPrincipalVariation[0];
+        return State.Global.Table.SavedPrincipalVariation[0];
     }
     
     for (Ply depth = 2; depth < Constants::MaxDepth; depth++)
     {
-        score = Aspiration(board, data, depth, score);
+        score = Aspiration(board, depth, score);
         callbackData.Depth = depth;
         callbackData._Score = score;
 
-        if(data.Stopper.ShouldStopDepthIncrease())
+        if(State.Stopper.ShouldStopDepthIncrease())
         {
             break;
         }
 
-        data.Global.Table.SavePrincipalVariation(board);
-        callback(callbackData);
+        State.Global.Table.SavePrincipalVariation(board);
+        Callback(callbackData);
     }
 
-    return data.Global.Table.SavedPrincipalVariation[0];
+    return State.Global.Table.SavedPrincipalVariation[0];
 }
 
-Move Search::Run(Board& board, const SearchParameters& parameters, const SearchCallback& callback)
+Move Search::Run(Board& board, const SearchParameters& parameters)
 {
-    const SearchStopper stopper = SearchStopper(parameters, board.WhiteToMove);
-    SearchData data = SearchData(stopper);
-    data.Global.Table.SetSize(1024 * 1024 * 1024);
-    const auto move = IterativeDeepen(board, data, callback);
+    State.Stopper.Init(parameters, board.WhiteToMove);
+    const auto move = IterativeDeepen(board);
     return move;
 }